+++ plugin.xml
@@ -48,8 +48,7 @@
         <header-file src="src/ios/Plugins/AppDelegate+JPush.h" />
         <source-file src="src/ios/Plugins/AppDelegate+JPush.m" />
 
-        <header-file src="src/ios/lib/JPUSHService.h" />
-        <source-file src="src/ios/lib/jpush-ios-4.8.1.a" framework="true" />
+        <source-file src="src/ios/lib/jpush-ios-5.2.0.xcframework" framework="true" />
         <resource-file src="src/ios/JPushConfig.plist" />
 
         <framework src="CFNetwork.framework" weak="true" />
